requirements.txt
click~=7.1.2
PyYAML~=5.3.1
rich~=6.0.0
